Make api_crawler a library module.
